restructure with screens and components folder
